v24: minimal covenants + chipnet re-genesis; two structural size limits closed
Ships the full v24 line and re-genesises chipnet. v24 began as Phase-2 daemon
hardening + Phase-1 covenant hardening, but deploying it surfaced TWO hard
structural limits that the audit-driven additions had blown past. Both are now
fixed by shrinking the covenants to their minimal trust-critical core (the
removed gates were redundant under the closed-category invariant) and guarded
by CI + deploy pre-flight so they can't silently recur.

THE TWO WALLS (caught empirically: a node REJECT_NONSTANDARD, then a libauth
VM trace):
  1. P2S locking_bytecode ≤ 201 B (relay policy). The PublisherSlot is P2S;
     v24's first cut reached 233 B → slot genesis rejected REJECT_NONSTANDARD.
  2. P2SH-32 redeem ≤ 520 B (consensus MAX_SCRIPT_ELEMENT_SIZE — the redeem is
     pushed as one stack item on spend). The Oracle is P2SH-32; v24's first cut
     reached a 537 B redeem → every Oracle.update failed "stack item exceeds
     520 bytes". cargo + mem-cash cannot see either limit (script-VM semantics
     only); only the node / a VM element-size check does.

THE FIX — minimal-slot / minimal-Oracle. Slot bytes cost ~26×/cycle (emitted
into 26 outputs: 13 attests + 13 re-emits); Oracle redeem is revealed 1×/cycle.
The closed-category invariant (genesis mints exactly 13 mutable slot NFTs, no
minting token ⇒ category permanently closed) makes the v24-first-cut additions
redundant, so they are dropped rather than relocated:
  * Slot 233 → 167 B: drop pkh-monotone sort (never gave independent security
    over F01 — V22-OC-22 proved it), u40 caps (cycleSeq self-bounded by the
    monotonic check; timestamp threshold-absorbed), length pins (self-pinned
    re-emit), tokenAmount pins (no category fungibles can exist). 167 B is
    *smaller* than the live v23 (196 B) — a fee improvement.
  * Oracle 501 → 461 B (redeem 537 → 497): drop the same class of redundant
    pins/caps; keep the u40 commit WIDTH (the actual F-OC12 Y2038 close) and F01.
  KEPT across both: Y2038 u40 widening (18-B Oracle/slot, 19-B Ticker), F01
  slot-category pin, all median/quorum/identity/sig/re-emit gates.

PERMANENT GUARDS (the bug class can't recur silently):
  * cargo: slot_template_fits_standardness_cap (≤201), oracle_redeem_fits_
    p2sh_element_cap + ticker_redeem_fits_p2sh_element_cap (≤520).
  * ops/deploy.rs pre-flight: refuses genesis if any slot body >201 B or any
    P2SH redeem >520 B.
  Also fixed a latent genesis bug: Oracle/slot initial commits were 16 B but
  v24 commits are 18 B (would have bricked the first cycle).

Phase-2 daemon hardening retained (P04 panic=unwind + single-exit + core-dump
suppression; P05 manifest TOFU hash-pin + 13/13 cross-daemon LB verify; P06
slot/Oracle template fingerprint CI gate; P07 systemd hardening + F13 manifest
perm gate). Docs: docs/design/minimal-slot-principle.md (the 26× rule, the
closed-category invariant, the 201/520 walls, the P2SH-32 escape hatch).

Workspace tests 141/141. Empirically validated on our BCHN node:
testmempoolaccept allowed both genesis txs; libauth VM trace pinned the 537-B
Oracle redeem failure; post-fix the live 13-daemon fleet cycles cleanly
(Oracle.update ok, seq advancing, median ~$199.6, zero rejections).

Chipnet re-genesis 2026-06-11:
  Oracle  cat 9c18d35925fd5a4a9af274b2a0555a83efec7488006fdf8db3a20ee6d3ec54ab
  Slot    cat ad44838456f7a026a46de4fe6025b30194dc0773b051409e53258e0ab001555f
  Oracle  addr bchtest:pvgts6s4qkmpjsalyqmc02dqu7zce0kfszkct2cssewhj0hvgtlrqhfqpj7q0
  web client re-pinned + u40 (18-B) commit decoders.

+++ node/core/src/cycle/error.rs
@@ -24,9 +24,9 @@ pub enum CycleError {
     #[error("my slot not found in {count} mutable slot UTXOs")]
     MySlotNotFound { count: usize },
     #[error("slot already at cycleSeq {at}, can't progress to {new}")]
-    SlotAheadOfNew { at: u32, new: u32 },
+    SlotAheadOfNew { at: u64, new: u64 },
     #[error("stride floor: waiting {wait_sec} s")]
-    StrideFloor { wait_sec: u32 },
+    StrideFloor { wait_sec: u64 },
     #[error("price fetch failed for source {source_id}: {reason}")]
     PriceFetchFailed { source_id: u16, reason: String },
     #[error("insufficient funder balance {have} < need {need}")]
